Reworking test structure - Automatic assignment of default variables - Allowing only ONE template per Makefile now
Lots of changes - should have been seperated, but that did not work out.

Tests: are an integral part  of the Makefiles now. run 'moa unittest' or 'moa unittest_base' in any moa directory - this means that the somewhat messy test structure is slowly being removed.

Default variables (defined as VARNAME_default) are now automatically assigned if they are undefined. A Todo here is variable_type checking

The biggest change is that only ONE template is allowed now PER MOA Makefile. order of including is very important in Makefiles and things were getting hairy. So. Now it is one makefile per template.
Combining makefiles is still possible, but has to be done at the template level. It was reather complicated anyhow.